+++ R/LearnerRegrCVGlmnet.R
@@ -12,7 +12,7 @@
 #' @template section_dictionary_learner
 #'
 #' @references
-#' \cite{mlr3learners}{friedman_2010}
+#' `r tools::toRd(bibentries["friedman_2010"])`
 #'
 #' @export
 #' @template seealso_learner