+++ R/LearnerSurvXgboost.R
@@ -30,7 +30,7 @@
 #' @templateVar id surv.xgboost
 #'
 #' @references
-#' \cite{mlr3learners}{chen_2016}
+#' `r tools::toRd(bibentries["chen_2016"])`
 #'
 #' @export
 #' @template seealso_learner