+++ R/helpers.R
@@ -6,13 +6,12 @@ pvec2mat = function(p, levs) {
   y
 }
 
-opts_default_contrasts = list(contrasts = c("contr.treatment", "contr.poly"))
-
 
 ordered_features = function(task, feature_names) {
   fn = intersect(feature_names, task$feature_names)
   task$data(cols = fn)
 }
+
 
 glmnet_feature_names = function(model) {
   beta = model$beta
@@ -20,3 +19,6 @@ glmnet_feature_names = function(model) {
     beta = model$glmnet.fit$beta
   rownames(if (is.list(beta)) beta[[1L]] else beta)
 }
+
+
+opts_default_contrasts = list(contrasts = c("contr.treatment", "contr.poly"))