+++ R/zzz.R
@@ -5,13 +5,10 @@
 #' @importFrom mlr3 mlr_learners LearnerClassif LearnerRegr
 #'
 #' @description
-#' More learners are available in the `mlr3learners` repository on Github
-#' (\url{https://github.com/mlr3learners}). There also is a wiki page listing
-#' all currently available custom learners
-#' (\url{https://github.com/mlr-org/mlr3learners/wiki/Extra-Learners}). A guide
-#' on how to create custom learners is covered in the book:
-#' \url{https://mlr3book.mlr-org.com}. Feel invited to contribute a missing
-#' learner to the \CRANpkg{mlr3} ecosystem!
+#' More learners are implemented in the [mlr3extralearners package](https://github.com/mlr-org/mlr3extralearners).
+#' A guide on how to create custom learners is covered in the book:
+#' \url{https://mlr3book.mlr-org.com}.
+#' Feel invited to contribute a missing learner to the \CRANpkg{mlr3} ecosystem!
 "_PACKAGE"
 
 register_mlr3 = function() {