+++ inst/paramtest/test_paramtest_regr.cv_glmnet.R
@@ -1,5 +1,7 @@
 library(mlr3learners)
 
+skip_on_os("solaris")
+
 test_that("regr.cv_glmnet", {
   learner = lrn("regr.cv_glmnet")
   fun = glmnet::cv.glmnet
@@ -12,7 +14,7 @@ test_that("regr.cv_glmnet", {
   ParamTest = run_paramtest(learner, fun, exclude)
   expect_true(ParamTest, info = paste0(
     "Missing parameters:",
-    paste0("- '", ParamTest$missing, "'", collapse = "‚")))
+    paste0("- '", ParamTest$missing, "'", collapse = ",")))
 })
 
 # example for checking a "control" function of a learner
@@ -27,7 +29,7 @@ test_that("regr.cv_glmnet", {
   ParamTest = run_paramtest(learner, fun, exclude)
   expect_true(ParamTest, info = paste0(
     "Missing parameters:",
-    paste0("- '", ParamTest$missing, "'", collapse = "‚")))
+    paste0("- '", ParamTest$missing, "'", collapse = ",")))
 })
 
 test_that("predict regr.cv_glmnet", {
@@ -41,5 +43,5 @@ test_that("predict regr.cv_glmnet", {
   ParamTest = run_paramtest(learner, fun, exclude)
   expect_true(ParamTest, info = paste0(
     "Missing parameters:",
-    paste0("- '", ParamTest$missing, "'", collapse = "‚")))
+    paste0("- '", ParamTest$missing, "'", collapse = ",")))
 })