+++ tests/testthat/test_classif_cv_glmnet.R
@@ -1,5 +1,6 @@
 context("classif.cv_glmnet")
 
+skip_on_os("solaris")
 skip_if_not_installed("glmnet")
 
 test_that("autotest", {